refactor: Reorganize validation components into plugins module
Move ValidatorFactory and CategoryValidator from validator.py to plugins.py
to better organize the codebase and prepare for future plugin architecture.

### 🔧 Core Changes:

**File Operations:**
- Created new sloth/plugins.py module with validation classes
- Updated imports across 9 files to use plugins module instead of validator

**Updated Import Statements in:**
- sloth/__init__.py: Main package imports
- sloth/common.py: Base class imports
- sloth/handler.py: Handler class imports
- sloth/legacy/parser.py: Legacy parser imports
- sloth/loaders.py: Data loader imports
- sloth/models.py: Core model imports
- sloth/parser.py: Main parser imports
- sloth/pdbml_converter.py: XML converter imports (variable rename)
- tests/test_schemas.py: Test fixture updates

**Classes Moved:**
- ValidatorFactory: Factory for creating validators and cross-checkers
- CategoryValidator: Validation executor for mmCIF categories

### 📂 Files Modified: 10 files
- 1 new file: sloth/plugins.py (89 lines)
- 9 files updated: Import statements and variable renames

### 🎯 Benefits:

1. **Better Organization**: Validation logic consolidated in plugins module
2. **Future-Ready**: Prepares foundation for extensible plugin architecture
3. **Cleaner Separation**: Separates core validation from other components
4. **Maintainability**: Centralized location for all plugin-related functionality

### ✅ Quality Assurance:

- All imports updated consistently across codebase
- No functional changes - pure refactoring
- Maintains full backwards compatibility
- Test naming updated to reflect new structure

This refactoring sets the foundation for a more modular, plugin-based
architecture while maintaining all existing validation functionality.
